requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
djongo==1.3.6
idna==3.3
mongoengine==0.23.1
oauthlib==3.2.0
pymongo==3.12.3
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.2.4
urllib3==1.26.8
